requirements.txt
python-dotenv>=1.1.1
rich>=14.0.0
